[424131] JSDT Parser doesn't respect "Enable Semantic Validation" preference value
Problem reported is made to respect Enable Semantic Validation preference value for the preferences, so the problems market as
ProblemSeverities.Optional (configurable problems) aren't reported in case of Semantic Validation is disabled.
With this fix the build validation and as-you-type validation (errpr/warning annotations) are showing the same results,
not configurable error/warning messages do not depend on the value of "Enable JavaScript semantic validation" preference value.
All the JUnit Test cases of DualParseSyntaxErrorTest and SyntaxErrorTest passed successfully.

Signed-off-by: vrubezhny <vrubezhny@exadel.com>